Reorganized types and added contract state typings.
