+++ ui/components/CampaignUtils.tsx
@@ -10,27 +10,27 @@ import { Button, StatusIndicator } from "@/components"
 import { payTokenSymbol } from "@/config"
 import { prettyPrintDecimal } from "@/helpers"
 import { favoriteCampaignAddressesAtom } from "@/state"
-import { CampaignActionType, Color, Status } from "@/types"
+import { CampaignActionType, CampaignStatus, Color } from "@/types"
 
-export const CampaignStatus: FunctionComponent<CampaignProps> = ({
+export const CampaignStatusIndicator: FunctionComponent<CampaignProps> = ({
   campaign: { status },
   className,
 }) => {
   let color: Color = Color.Placeholder
   let label: string
   switch (status) {
-    case Status.Pending:
+    case CampaignStatus.Pending:
       label = "Pending"
       break
-    case Status.Open:
+    case CampaignStatus.Open:
       color = Color.Orange
       label = "Open"
       break
-    case Status.Funded:
+    case CampaignStatus.Funded:
       color = Color.Green
       label = "Funded"
       break
-    case Status.Cancelled:
+    case CampaignStatus.Cancelled:
       label = "Cancelled"
       break
     default:
@@ -62,7 +62,8 @@ export const CampaignProgress: FunctionComponent<CampaignProgressProps> = ({
 }) => {
   // Round down so we don't say 100% funded until it has actually been funded.
   const fundedPercent = Math.floor((100 * pledged) / goal)
-  const showProgress = status === Status.Open || status === Status.Funded
+  const showProgress =
+    status === CampaignStatus.Open || status === CampaignStatus.Funded
 
   return (
     <div className={cn("flex flex-col justify-start w-full", className)}>
