+++ ui/types/keplr.ts
@@ -1,0 +1,5 @@
+import { Window as KeplrWindow } from "@keplr-wallet/types"
+
+declare global {
+  interface Window extends KeplrWindow {}
+}
